+++ FICS/utils.h
@@ -31,6 +31,7 @@
 #ifndef _UTILS_H
 #define _UTILS_H
 
+#include <stdint.h>
 #include <stdio.h>
 
 #include "common.h" /* PRINTFLIKE() */
@@ -104,7 +105,7 @@ extern int		 safechar(int);
 extern int		 safestring(char *);
 extern int		 search_directory(char *, char *, char **, int);
 extern int		 truncate_file(char *, int);
-extern time_t		 untenths(unsigned int);
+extern time_t		 untenths(uint64_t);
 extern unsigned int	 tenth_secs(void);
 //extern void		 pprintf_dohightlight(int);
 //extern void		 sprintf_dohightlight(int, char *);
